backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
openai
httpx
python-multipart
mcp
mcp
psycopg2-binary
supabase
